requirements.txt
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
openai>=1.30.0
pytest>=8.0.0
